+++ src/Microweber/Application.php
@@ -39,6 +39,7 @@ namespace Microweber;
  * @property \Microweber\Utils\Captcha                                     $captcha
  * @property \Microweber\Providers\FormsManager                            $forms_manager
  * @property \Microweber\Providers\Content\DataFieldsManager               $data_fields_manager
+ * @property \Microweber\Providers\Content\AttributesManager               $attributes_manager
  *
  */
 class Application {