+++ src/Microweber/Providers/ContentManager.php
@@ -552,47 +552,13 @@ class ContentManager {
     }
 
     public function attributes($content_id) {
-
+        
         $data = array();
         $data['rel_type'] = 'content';
         $data['rel_id'] = intval($content_id);
-        $res = array();
-        $get = $this->get_attributes($data);
-        if (!empty($get)){
-            foreach ($get as $item) {
-                if (isset($item['attribute_name']) and isset($item['attribute_value'])){
-                    $res[ $item['attribute_name'] ] = $item['attribute_value'];
-                }
-            }
-        }
-        if (!empty($res)){
-            return $res;
-        }
-
-        return $get;
+        return $this->app->attributes_manager->get_values($data);
     }
 
-    public function get_attributes($data = false) {
-        $table = $this->tables['attributes'];
-        if (is_string($data)){
-            $data = parse_params($data);
-        }
-        if (!is_array($data)){
-            $data = array();
-        }
-        $data['table'] = $table;
-        $get = $this->app->database_manager->get($data);
-        if (!empty($get)){
-            foreach ($get as $k => $data) {
-                if (isset($data['attribute_value']) and isset($data['attribute_type']) and ($data['attribute_type']=='array')){
-                    $data['attribute_value'] = json_decode($data['attribute_value'], true);
-                    $get[ $k ] = $data;
-                }
-            }
-        }
-
-        return $get;
-    }
 
     /**
      * paging
@@ -4374,7 +4340,6 @@ class ContentManager {
         return $this->app->data_fields_manager->save($data);
 
 
-
         $table = $this->tables['content_data'];
         if (!is_array($data)){
             $data = parse_params($data);
@@ -4424,56 +4389,6 @@ class ContentManager {
         return $save;
     }
 
-
-    public function save_content_attribute($data, $delete_the_cache = true) {
-        $table = $this->tables['attributes'];
-        if (!is_array($data)){
-            $data = parse_params($data);
-        }
-        if (!isset($data['id'])){
-            if (!isset($data['attribute_name'])){
-                return array('error' => "You must set 'field' parameter");
-            }
-            if (!isset($data['attribute_value'])){
-                return array('error' => "You must set 'value' parameter");
-            }
-        }
-        if (!isset($data['rel_type']) and isset($data['content_id'])){
-            $data['rel_type'] = 'content';
-            $data['rel_id'] = $data['content_id'];
-        }
-        if (isset($data['attribute_name']) and isset($data['rel_id']) and isset($data['rel_type'])){
-            $is_existing_data = array();
-            $is_existing_data['attribute_name'] = $data['attribute_name'];
-            $is_existing_data['rel_id'] = $data['rel_id'];
-            $is_existing_data['rel_type'] = $data['rel_type'];
-            $is_existing_data['one'] = true;
-            $is_existing = $this->get_attributes($is_existing_data);
-            if (is_array($is_existing) and isset($is_existing['id'])){
-                $data['id'] = $is_existing['id'];
-            }
-        }
-        if (isset($data['content_id'])){
-            $data['rel_id'] = intval($data['content_id']);
-        }
-        if (isset($data['attribute_value']) and is_array($data['attribute_value'])){
-            $data['attribute_value'] = json_encode($data['attribute_value']);
-            $data['attribute_type'] = 'array';
-        }
-        if (!isset($data['rel_type'])){
-            $data['rel_type'] = 'content';
-        }
-
-        if (isset($data['rel_type']) and $data['rel_type']=='content'){
-            if (isset($data['rel_id'])){
-                $data['content_id'] = $data['rel_id'];
-            }
-        }
-        $save = $this->app->database_manager->save($table, $data);
-        $this->app->cache_manager->delete('attributes');
-
-        return $save;
-    }
 
     function create_default_content($what) {